Added coding exercise gist.
This is a draft of a Pythonic template. It is supposed to follow PEP-8 style guidelines, but I haven't yet decided how to fully encapsulate this.

If you'd like to practice coding challenges and thinking through a problem, I recommend using HackerRank or InterviewCake.

If you're new to coding, I say take a look at freeCodeCamp.org.